bot/requirements.txt
aiogram==3.4.1
pillow==10.4.0
opencv-python==4.10.0.84
requests==2.32.3
numpy==1.26.4
python-dotenv==1.0.1
